requirements.txt
regex>=2023.12.25
rank-bm25>=0.2.2
numpy>=1.24,<2.0
tqdm>=4.66.0
python-dotenv>=1.0.1
openai>=1.40.0
sentence-transformers>=2.7.0


